TPASPKT-1033 Write METS and files to SIP
Add functionality to write the given METS file as well as the files to
given file path.

METS is built as usual, but using SIPDigitalObjects from this project
rather than mets_builder DigitalObjects. The reasoning is that
SIPDigitalObject can have attributes and functionalities that do not
make sense in the mets_builder context, for now meaning the local file
path of the file. In the future, SIPDigitalObject could potentially have
functions that for example validate the file or generate metadata by
scraping the file (generating the corresponding mets_builder metadata
objects and adding the metadata object to itself automatically).

Files are then copied to the SIP as they are declared in the METS
object.