TPASPKT-1047 Refactor digital object source path validation
Make a setter for source filepath, to ensure that validations are run
even if the user changes the filepath after initializing the object.

Additionally clarifies why Path.resolve is run with a comment.

Refactoring the code caused error message to spell the entire absolute
resolved path of the source file instead of jsut the relative path, if
that is what the user gave. This demanded some changes to the test that
asserts the error text.