TPASPKT-1223 Generate container bitstream metadata
Generate metadata objects for individual bitstreams if the parsed file is
a container. The bitstream and media metadata objects are linked to the
container metadata object.

For example, a Matroska video file might contain video and/or audio
streams; the Matroska container is represented using a
`TechnicalFileObjectMetadata`, the video and audio bitstreams are
represented using `TechnicalBitstreamObjectMetadata` and the audio/video
metadata per stream is represented using
`Technical{Audio,Video}Metadata`. Each stream in METS `stream` object
contains links to both the bitstream and media metadata.