TPASPKT-1280 Add method for creating siptools-ng agent
As agent metadata representing dpres-siptools-ng would be commonly
needed, add a convenience method to create that easily.

Hard coded premis id (which is separate from the METS metadata id) for
the agent is added to help with deduplicating the metadata: a random
generated premis id would lead to unidentical metadata instances each
time the method is called. With a constant premis id separate calls to
the method lead to identical instances, and the mets-builder serializer
can handle the metadata deduplication automatically.