TPASPKT-1225 Add method for creating file-scraper agent
Agent metadata representing file-scraper will be needed to be added as
agent for certain events, so add a convenience method to create that
easily.

Hard coded premis id (which is separate from the METS metadata id) for
the agent is added to help with deduplicating the metadata: a random
generated premis id would lead to unidentical metadata instances each
time the method is called. With a constant premis id separate calls to
the method lead to identical instances, and the mets-builder serializer
can handle the metadata deduplication automatically.