TPASPKT-1225 Add checksum calculation event
Add event documenting checksum calculation for digital objects. The
event is not added when the user provides their own checksum when
generating metadata for the digital object.

The event is implemented differently than in old siptools: here the
event is added as metadata to each digital object for which the checksum
was calculated for, instead of adding the event as metadata to the root
div of a structural map (implying that calculation was done for every
digital object). Adding the event as metadata to each digital object
individually makes more sense for siptools-ng architecture (a natural
place to create the event is in generate_technical_metadata, when
creating metadata for single digital object) and is also more exact, as
the event is applied directly to each digital object and omitted from
digital objects for which the checksum was not calculated for.

Some tests that previously tested that digital objects have certain
amount of metadata were modified, since adding the checksum calculation
event and agents change the metadata object count. As these kind of
changes might happen in the future as well, the metadata count is not
tested anymore at all.