TPASPKT-1375 Add `SIP.from_files`
Add `SIP.from_files` class method as a convenience function for
generating a SIP from File instances. This lets the developer better
organize the directory hierarchy in the SIP and add additional metadata to
the digital objects before letting siptools-ng handle the rest, giving
more control for the user compared to `SIP.from_directory`.

The previous implementation of `SIP` constructor automatically generated
technical metadata (which caused a crash if the user had already done
this themselves), as well generated a structural map for the user. The
constructor now only sets the fields; the docstring is updated to direct
users towards `SIP.from_directory` and `SIP.from_files` depending on
which is more appropriate for their usage.