TPASPKT-1452 Skip tech metadata for bit-level
Skip generation of specific technical metadata objects (eg. audio,
video) when importing a bit-level file. The generated metadata might be
incorrect or incomplete for bit-level files, meaning it's not reasonable
to add them to the resulting METS document. In some cases, such as the
ProRes file, the technical metadata generation will crash due to the
resulting element not being compliant with the spec.